backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
xgboost==2.1.1
onnxruntime==1.24.3
networkx==3.3
numpy>=2.0.0
pydantic==2.8.2
firebase-admin==6.5.0
python-dotenv==1.0.1
apscheduler==3.10.4
pytz==2024.1
scikit-learn==1.5.2
pandas==2.2.2
httpx==0.27.2
